+++ programs/drift/src/controller/orders.rs
@@ -1031,7 +1031,9 @@ pub fn fill_perp_order(
         return Ok(0);
     }
 
-    let should_expire_order = should_expire_order(user, order_index, now)?;
+    validate_perp_fill_possible(state, user, order_index, slot, makers_and_referrer.0.len())?;
+
+    let should_expire_order = should_expire_order_before_fill(user, order_index, now)?;
 
     let position_index =
         get_position_index(&user.perp_positions, user.orders[order_index].market_index)?;
@@ -3216,7 +3218,7 @@ pub fn fill_spot_order(
         }
     }
 
-    let should_expire_order = should_expire_order(user, order_index, now)?;
+    let should_expire_order = should_expire_order_before_fill(user, order_index, now)?;
 
     let should_cancel_reduce_only = if user.orders[order_index].reduce_only {
         let market_index = user.orders[order_index].market_index;